move the log accumulation to post executions
